Still developing dcmsearch:
cm_pipeline.c: fixed bug, I wasn't reconfiguring non-optimized profile
	       for length of window 
cm_mx.c, structs.h: changed <ncells> counter of DP cells in CM_HB_MX data
	            structure from int to int64_t
cm_dpsearch.c: cosmetic changes
cm_p7_domaindef.c: minor changes
dcmsearch.c: added some experimental and probably temporary options


git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/trunk@3390 db3e70e4-12e1-0310-b23f-81a6e3d09ba7

+++ src/cm_mx.c
@@ -135,9 +135,9 @@ cm_hb_mx_GrowTo(CM_t *cm, CM_HB_MX *mx, char *errbuf, CP9Bands_t *cp9b, int L, f
   void   *p;
   int     v, jp;
   int     cur_size = 0;
-  size_t  ncells;
+  int64_t ncells;
   int     jbw;
-  double  Mb_needed;
+  float   Mb_needed;
   int     have_el;
   have_el = (cm->flags & CMH_LOCAL_END) ? TRUE : FALSE;
 
@@ -146,7 +146,7 @@ cm_hb_mx_GrowTo(CM_t *cm, CM_HB_MX *mx, char *errbuf, CP9Bands_t *cp9b, int L, f
    * cp9b has */
   if(cp9b == NULL)        ESL_FAIL(eslEINCOMPAT, errbuf, "cm_hb_mx_GrowTo() entered with cp9b == NULL.\n");
   if(cp9b->cm_M != mx->M) ESL_FAIL(eslEINCOMPAT, errbuf, "cm_hb_mx_GrowTo() entered with mx->M: (%d) != cp9b->M (%d)\n", mx->M, cp9b->cm_M);
-  
+
   ncells = 0;
   Mb_needed = ((float) (sizeof(int *)) * ((float) mx->M + 1)) + /* nrowsA ptrs */
     (float) (sizeof(float **)) * (float) (mx->M);               /* mx->dp[] ptrs */
